+++ test/functional/abc_p2p_avalanche_transaction_finalization.py
@@ -220,6 +220,7 @@ class AvalancheTransactionFinalizationTest(BitcoinTestFramework):
 
         assert_equal(node.getmempoolinfo()["size"], num_txs)
         assert_equal(node.getmempoolinfo()["bytes"], 13000)
+        assert_equal(node.getmempoolinfo()["finalized_txs_size"], num_txs)
         assert_equal(node.getmempoolinfo()["finalized_txs_bytes"], 13000)
 
         # Lets build a chain of 8 txs of size 200 each. Each individual tx would
@@ -230,6 +231,7 @@ class AvalancheTransactionFinalizationTest(BitcoinTestFramework):
         )
         assert_equal(node.getmempoolinfo()["size"], num_txs + 8)
         assert_equal(node.getmempoolinfo()["bytes"], 14600)
+        assert_equal(node.getmempoolinfo()["finalized_txs_size"], num_txs)
         assert_equal(node.getmempoolinfo()["finalized_txs_bytes"], 13000)
 
         # We can finalize the first 4 txs but not the last 4
@@ -243,6 +245,7 @@ class AvalancheTransactionFinalizationTest(BitcoinTestFramework):
 
         assert_equal(node.getmempoolinfo()["size"], num_txs + 8)
         assert_equal(node.getmempoolinfo()["bytes"], 14600)
+        assert_equal(node.getmempoolinfo()["finalized_txs_size"], num_txs + 4)
         assert_equal(node.getmempoolinfo()["finalized_txs_bytes"], 13800)
 
         # Add another tx to the chain. This tx size is 100 bytes, so it gets
@@ -272,6 +275,7 @@ class AvalancheTransactionFinalizationTest(BitcoinTestFramework):
 
         assert_equal(node.getmempoolinfo()["size"], num_txs + 9)
         assert_equal(node.getmempoolinfo()["bytes"], 14700)
+        assert_equal(node.getmempoolinfo()["finalized_txs_size"], num_txs + 4)
         assert_equal(node.getmempoolinfo()["finalized_txs_bytes"], 13800)
 
         # Add a transaction that will not get polled because it doesn't fit
@@ -282,6 +286,7 @@ class AvalancheTransactionFinalizationTest(BitcoinTestFramework):
         assert big_tx["txid"] in node.getrawmempool()
         assert_equal(node.getmempoolinfo()["size"], num_txs + 10)
         assert_equal(node.getmempoolinfo()["bytes"], 16700)
+        assert_equal(node.getmempoolinfo()["finalized_txs_size"], num_txs + 4)
         assert_equal(node.getmempoolinfo()["finalized_txs_bytes"], 13800)
 
         # This tx will not be polled. We can check this by looking at the vote
@@ -317,6 +322,7 @@ class AvalancheTransactionFinalizationTest(BitcoinTestFramework):
 
         assert_equal(node.getmempoolinfo()["size"], 6)
         assert_equal(node.getmempoolinfo()["bytes"], 4 * 200 + 100 + 2000)
+        assert_equal(node.getmempoolinfo()["finalized_txs_size"], 0)
         assert_equal(node.getmempoolinfo()["finalized_txs_bytes"], 0)
 
         mempool = node.getrawmempool()
@@ -331,6 +337,7 @@ class AvalancheTransactionFinalizationTest(BitcoinTestFramework):
 
         assert_equal(node.getmempoolinfo()["size"], 6)
         assert_equal(node.getmempoolinfo()["bytes"], 4 * 200 + 100 + 2000)
+        assert_equal(node.getmempoolinfo()["finalized_txs_size"], 6)
         assert_equal(node.getmempoolinfo()["finalized_txs_bytes"], 4 * 200 + 100 + 2000)
 
     def test_blockmintxfee(self):