+++ doc/release-notes.md
@@ -4,4 +4,6 @@ Bitcoin ABC version 0.32.11 is now available from:
 
   <https://download.bitcoinabc.org/0.32.11/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+ - The `getmempoolinfo` RPC now includes a `finalized_txs_size` with the count
+   finalized transactions.